requirements.txt
aiohttp==3.13.3
beautifulsoup4==4.14.3
g4f==7.2.5
gspread==6.2.1
howlongtobeatpy==1.0.21
oauth2client==4.1.3
openai==2.24.0
python-dotenv==1.2.1
SQLAlchemy==2.0.48
twitchio==2.8.2
